+++ org.go
@@ -3,9 +3,9 @@ package podio
 import "fmt"
 
 type Organization struct {
-	Id   uint   `json:"org_id"`
-	Slug string `json:"url_label"`
-	Name string `json:"name"`
+	Id   float64 `json:"org_id"`
+	Slug string  `json:"url_label"`
+	Name string  `json:"name"`
 }
 
 func (client *Client) GetOrganizations() (orgs []Organization, err error) {
@@ -13,8 +13,8 @@ func (client *Client) GetOrganizations() (orgs []Organization, err error) {
 	return
 }
 
-func (client *Client) GetOrganization(id uint) (org *Organization, err error) {
-	path := fmt.Sprintf("/org/%d", id)
+func (client *Client) GetOrganization(id float64) (org *Organization, err error) {
+	path := fmt.Sprintf("/org/%.0f", id)
 	err = client.Request("GET", path, nil, nil, &org)
 	return
 }
